+++ app/Orm/LoansRepository.php
@@ -40,7 +40,7 @@ final class LoansRepository extends \Nextras\Orm\Repository\Repository {
   public function findReturnedThisMonth(int $user): ICollection {
     $month = (int) date("n");
     $year = (int) date("Y");
-    $startOfMonthTS = mktime(0, 0, 0, $month, 1, $year);
+    $startOfMonthTS = (int) mktime(0, 0, 0, $month, 1, $year);
     $date = new \DateTime();
     $date->setTimestamp($startOfMonthTS);
     $start = $date->getTimestamp();
