+++ app/Orm/UserAdventuresRepository.php
@@ -50,7 +50,7 @@ final class UserAdventuresRepository extends \Nextras\Orm\Repository\Repository 
    * @return ICollection|UserAdventure[]
    */
   public function findFromMonth(int $user, int $month = null, int $year = null): ICollection {
-    $startOfMonthTS = mktime(0, 0, 0, $month ?? (int) date("n"), 1, $year ?? (int) date("Y"));
+    $startOfMonthTS = (int) mktime(0, 0, 0, $month ?? (int) date("n"), 1, $year ?? (int) date("Y"));
     $date = new \DateTime();
     $date->setTimestamp($startOfMonthTS);
     $start = $date->getTimestamp();
